app/requirements.txt
black==23.11.0
isort==5.12.0
jupyter==1.0.0
loguru==0.7.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
streamlit==1.28.2
streamlit-js-eval==0.1.5
python-dotenv==1.0.0